refactor: move type-related methods to Types
